Add student to all class assignments when join class.

See merge request upchieve/subway!1724

+++ server/models/Teacher/teacher.sql
@@ -74,7 +74,7 @@ SELECT
         FROM
             student_classes
         WHERE
-            class_id = id) AS total_students
+            class_id = id)::int AS total_students
 FROM
     teacher_classes
 WHERE
